[psb] fix USB full speed / high speed handling
This should fix spectrum acquisition for USB4000 and HR4000.
Note that in general USB1.0 full speed mode is likely to be broken
for all FPGA OOI spectrometers. This adds a warning in case someone
runs into the issue. Implementing full speed mode for those
spectrometers requires someone with the hardware to test this.

The warning is emitted in case a spectrometer is forced to run in
full speed mode and recommends reporting it to upstream. My guess
is that there are not many USB1.0 hosts out there anymore, but
who knows, maybe there's some old lab PC that is crucial to some
experiment.

NOTE: A USB1.0 or USB1.1 hub would allow testing the spectrometers
in FS mode.

Close #87. Close #72.