[psb] support dynamic class substitution for flame-s
Allow a FlameS spectrometer to be dynamically converted from a
USB2000Plus. I currently believe that because USB2000+ and FlameS
share the same USB ProductIds the way to detect a FlameS is that
its major firmware version stored on the fpga is >= 3.

This might be wrong. But I don't have the hardware nor the time
to investigate this further. And my USB2000+ has firmware version
1.x. And I know of one FlameS with firmware version 3.x :)

See: #92