[pyseabreeze] add support for ADC1000-USB via USB
The ADC1000-USB is an older interface for S-series spectrometers (eg.
S200) and has USB and serial connections.

This patch adds support for the interfacing with it via USB.

There are some similar ADC cards using different transports (eg. PCI,
PCMCIA), which I suspect use the same protocol, so this patch may set
the infrastructure to support those too.

Signed-off-by: Filipe Laíns <lains@riseup.net>